moved thermal model data over to model controller
