+++ SolarThermalSystem/Simulator/SimulatorViewController.swift
@@ -19,11 +19,13 @@ class SimulatorViewController: UIViewController {
     @IBOutlet weak var hotWaterDemandTextField: UITextField!
     @IBOutlet weak var systemImageView: UIImageView!
     @IBOutlet weak var backgroundView: UIView!
-    
+    @IBOutlet weak var runSimulationButton: UIButton!
+    @IBOutlet weak var dailyHeatTransferLabel: UILabel!
+    @IBOutlet weak var solarInputLabel: UILabel!
     
     // MARK: - Properties
     // Initialize Model Properties
-    var thermalSystem = ThermalSystem.shared
+    var viewModel = SimulatorModelController.shared
     let dataSource = SystemDataSource()
     
     // MARK: - Lifecycle
@@ -52,14 +54,14 @@ class SimulatorViewController: UIViewController {
         setupCollectorOrientation()
         setupOvershading()
     }
-    
+
     // MARK: - Methods
     /// Each method that deals with user input values needs to conclude by assigning that value to the model variables that will be accessed in the "run simulator" function
     func setupCollectorType() {
         /// Collector type assigned will also determine the collectorEfficiency, the linear heat loss coefficient (used in function that assigns collectorPerformanceFactor), and ratio of aperture area to gross area (which will be multiplied by the dimensions provided by the user)
         ///
         let optionClosure = { (action: UIAction) in
-            self.thermalSystem.setCollectorType(with: action.title)
+            self.viewModel.setCollectorType(with: action.title)
             print(action.title)
         }
         var optionsArray = [UIAction]()
@@ -88,7 +90,7 @@ class SimulatorViewController: UIViewController {
     func setupCollectorTilt() {
         /// Tilt and orientation are used together in an array to determine the annual solar radiation kWh/m2
         let optionClosure = { (action: UIAction) in
-            self.thermalSystem.setCollectorTilt(with: action.title)
+            self.viewModel.setCollectorTilt(with: action.title)
             print(action.title)
         }
         var optionsArray = [UIAction]()
@@ -117,7 +119,7 @@ class SimulatorViewController: UIViewController {
     
     func setupCollectorOrientation() {
         let optionClosure = { (action: UIAction) in
-            self.thermalSystem.setCollectorOrientation(with: action.title)
+            self.viewModel.setCollectorOrientation(with: action.title)
             print(action.title)}
         
         var optionsArray = [UIAction]()
@@ -173,28 +175,9 @@ class SimulatorViewController: UIViewController {
     }
     
     // MARK: - Actions
-    
-    @IBAction func panelTypeChanged(_ sender: Any) {
-        guard let menuItem = collectorTypePopupButton.menu?.title else { return }
-        thermalSystem.setCollectorType(with: menuItem)
-    }
-    
-    @IBAction func tiltChanged(_ sender: Any) {
-        guard let menuItem = collectorTiltPopupButton.menu?.title else { return }
-        thermalSystem.setCollectorTilt(with: menuItem)
-    }
-    
-    @IBAction func orientationChanged(_ sender: Any) {
-        // assign vector of orientation and tilt to annual radiation
-    }
-    
-    @IBAction func overshadingChanged(_ sender: Any) {
-        // corresponding shading factor
-    }
-    
     @IBAction func runSimulationTapped(_ sender: Any) {
         // final daily heat transfer and solar input formulas
-        
+        // Case success, populate labels. Case failure, alert to check all fields for data.
     }
     
 } // End of Class
